🧑‍💻 (cli): Switch to a more generic cli tool
