+++ package.json
@@ -37,7 +37,6 @@
     "mobile": "pnpm --filter @ledgerhq/device-management-kit-mobile-sample",
     "doc": "pnpm --filter @ledgerhq/ledger-dmk-docs",
     "bump": "changeset version",
-    "prepare-release": "pnpm lmdk-release",
     "prerelease": "pnpm recursive exec -- pnpm pack",
     "release": "changeset publish",
     "changelog": "changeset add",
@@ -45,12 +44,13 @@
     "commitcl": "pnpm changelog && git add .changeset && pnpm commit",
     "danger:local": "pnpm danger local --dangerfile danger/dangerfile-local.ts --base origin/develop",
     "danger:local:fork": "pnpm danger local --dangerfile danger/dangerfile-local-fork.ts --base origin/develop",
-    "danger:ci": "pnpm danger ci --dangerfile danger/dangerfile-ci.ts --failOnErrors"
+    "danger:ci": "pnpm danger ci --dangerfile danger/dangerfile-ci.ts --failOnErrors",
+    "cli": "pnpm lmdk-cli"
   },
   "devDependencies": {
     "@changesets/changelog-github": "^0.5.0",
     "@changesets/cli": "^2.27.12",
-    "@ledgerhq/release-scripts": "workspace:*",
+    "@ledgerhq/ldmk-cli": "workspace:*",
     "@types/node": "^22.10.1",
     "@vitest/coverage-istanbul": "^3.0.4",
     "concurrently": "^9.1.0",
