+++ packages/mockserver-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ledgerhq/device-mockserver-client",
-  "private": true,
+  "private": false,
   "version": "1.1.0",
   "license": "Apache-2.0",
   "main": "lib/index.js",
